# perf: qualify statements without removing comments