--- a/google-cloud-spanner/src/main/java/com/google/cloud/spanner/connection/SimpleParser.java
+++ b/google-cloud-spanner/src/main/java/com/google/cloud/spanner/connection/SimpleParser.java
@@ -33,16 +33,23 @@ class SimpleParser {
    * if so, what the value was.
    */
   static class Result {
-    static final Result NOT_FOUND = new Result(null);
+    static final Result NOT_FOUND = new Result(null, false);
 
     static Result found(String value) {
-      return new Result(Preconditions.checkNotNull(value));
+      return new Result(Preconditions.checkNotNull(value), false);
+    }
+
+    static Result found(String value, boolean inParenthesis) {
+      return new Result(Preconditions.checkNotNull(value), inParenthesis);
     }
 
     private final String value;
 
-    private Result(String value) {
+    private final boolean inParenthesis;
+
+    private Result(String value, boolean inParenthesis) {
       this.value = value;
+      this.inParenthesis = inParenthesis;
     }
 
     @Override
@@ -55,7 +62,8 @@ class SimpleParser {
       if (!(o instanceof Result)) {
         return false;
       }
-      return Objects.equals(this.value, ((Result) o).value);
+      return Objects.equals(this.value, ((Result) o).value)
+          && Objects.equals(this.inParenthesis, ((Result) o).inParenthesis);
     }
 
     @Override
@@ -73,6 +81,10 @@ class SimpleParser {
     String getValue() {
       return this.value;
     }
+
+    boolean isInParenthesis() {
+      return this.inParenthesis;
+    }
   }
 
   // TODO: Replace this with a direct reference to the dialect, and move the isXYZSupported methods
@@ -80,6 +92,9 @@ class SimpleParser {
   private final AbstractStatementParser statementParser;
 
   private final String sql;
+
+  // TODO: Use this length field instead of repeatedly calling sql.length()
+  private final int length;
 
   private final boolean treatHintCommentsAsTokens;
 
@@ -100,6 +115,7 @@ class SimpleParser {
         !(treatHintCommentsAsTokens && dialect != Dialect.POSTGRESQL),
         "treatHintCommentsAsTokens can only be enabled for PostgreSQL");
     this.sql = sql;
+    this.length = sql.length();
     this.pos = pos;
     this.statementParser = AbstractStatementParser.getInstance(dialect);
     this.treatHintCommentsAsTokens = treatHintCommentsAsTokens;
@@ -117,10 +133,52 @@ class SimpleParser {
     return this.pos;
   }
 
+  void skipHint() {
+    // We don't need to do anything special for PostgreSQL, as hints in PostgreSQL are inside
+    // comments and comments are automatically skipped by all methods.
+    if (getDialect() == Dialect.GOOGLE_STANDARD_SQL && eatTokens('@', '{')) {
+      while (pos < length && !eatToken('}')) {
+        pos += statementParser.skip(sql, pos, /*result=*/ null);
+      }
+    }
+  }
+
+  Result eatNextKeyword() {
+    skipHint();
+    boolean inParenthesis = false;
+    while (pos < length && eatToken('(')) {
+      inParenthesis = true;
+    }
+    return eatKeyword(inParenthesis);
+  }
+
   /** Returns true if this parser has more tokens. Advances the position to the first next token. */
   boolean hasMoreTokens() {
     skipWhitespaces();
     return pos < sql.length();
+  }
+
+  /** Eats and returns the keyword at the current position. */
+  Result eatKeyword() {
+    return eatKeyword(false);
+  }
+
+  /**
+   * Eats and returns the keyword at the current position and returns a result that indicates that
+   * the keyword is inside one or more parentheses.
+   */
+  Result eatKeyword(boolean inParenthesis) {
+    if (!hasMoreTokens()) {
+      return Result.NOT_FOUND;
+    }
+    if (!Character.isLetter(sql.charAt(pos))) {
+      return Result.NOT_FOUND;
+    }
+    int startPos = pos;
+    while (pos < length && Character.isLetter(sql.charAt(pos))) {
+      pos++;
+    }
+    return Result.found(sql.substring(startPos, pos), inParenthesis);
   }
 
   /**