--- a/google-cloud-spanner/src/test/java/com/google/cloud/spanner/connection/ClientSideStatementsTest.java
+++ b/google-cloud-spanner/src/test/java/com/google/cloud/spanner/connection/ClientSideStatementsTest.java
@@ -18,6 +18,7 @@ package com.google.cloud.spanner.connection;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
@@ -62,6 +63,13 @@ public class ClientSideStatementsTest extends AbstractSqlScriptTest {
         throw SpannerExceptionFactory.newSpannerException(
             ErrorCode.INVALID_ARGUMENT, "Unknown or unsupported dialect: " + dialect);
     }
+  }
+
+  @Test
+  public void testIsQuery() {
+    AbstractStatementParser parser = AbstractStatementParser.getInstance(dialect);
+    ParsedStatement parsedStatement = parser.parse(Statement.of("show/spanner.statement_tag;"));
+    assertTrue(parsedStatement.isQuery());
   }
 
   @Test
@@ -293,7 +301,7 @@ public class ClientSideStatementsTest extends AbstractSqlScriptTest {
         log(
             statement.getExamplePrerequisiteStatements(),
             withInvalidSuffix(sql),
-            parser.isQuery(withInvalidSuffix(sql))
+            parser.parse(Statement.of(withInvalidSuffix(sql))).isQuery()
                 ? ErrorCode.UNIMPLEMENTED
                 : ErrorCode.INVALID_ARGUMENT);
       }
@@ -313,13 +321,13 @@ public class ClientSideStatementsTest extends AbstractSqlScriptTest {
           log(
               statement.getExamplePrerequisiteStatements(),
               withSuffix(replacement, sql),
-              parser.isQuery(withSuffix(replacement, sql))
+              parser.parse(Statement.of(withSuffix(replacement, sql))).isQuery()
                   ? ErrorCode.UNIMPLEMENTED
                   : ErrorCode.INVALID_ARGUMENT);
           log(
               statement.getExamplePrerequisiteStatements(),
               replaceLastSpaceWith(replacement, sql),
-              parser.isQuery(replaceLastSpaceWith(replacement, sql))
+              parser.parse(Statement.of(replaceLastSpaceWith(replacement, sql))).isQuery()
                   ? ErrorCode.UNIMPLEMENTED
                   : ErrorCode.INVALID_ARGUMENT);
         }