Determine the type of statement without first removing all comments and hints. This prevents the creation of new strings and stepping through the entire SQL string for each statement that is not found in the statement cache.

#### With change

```
Benchmark                                Mode  Cnt       Score      Error  Units
StatementParserBenchmark.isQueryTest    thrpt    5  547904.501 ± 1970.170  ops/s
StatementParserBenchmark.longDmlTest    thrpt    5  114806.782 ±  826.881  ops/s
StatementParserBenchmark.longQueryTest  thrpt    5  112666.992 ±  700.783  ops/s
```

#### Main

```
Benchmark                                Mode  Cnt       Score       Error  Units
StatementParserBenchmark.isQueryTest    thrpt    5  338079.227 ± 26995.410  ops/s
StatementParserBenchmark.longDmlTest    thrpt    5    1153.123 ±     5.604  ops/s
StatementParserBenchmark.longQueryTest  thrpt    5    2011.550 ±    72.433  ops/s
```

Updates #3800